Start of Rewrite
Differences substantial between current model voltage plots and Zhenyang's. As such I will be rewriting the methods of the OLM_SingleCompartment class to hopefully account for any errors made insofar.

Next Step is to properly implement RK4... maybe that will help? Further testing regarding the stimulation current will be tested, as spiking behaviour is unable to be produced in this model so far.